Change price storage

+++ db/migrate/20180216172238_create_snacks.rb
@@ -2,7 +2,7 @@ class CreateSnacks < ActiveRecord::Migration[5.1]
   def change
     create_table :snacks do |t|
       t.string :name
-      t.float :price
+      t.numeric :price
     end
   end
 end
